refactor(backend-requests): rename files to keep convention
